# Implement abstraction over mul_add