--- a/src/math.rs
+++ b/src/math.rs
@@ -177,6 +177,7 @@ pub fn expf(x: f32) -> f32 {
 
 /// Computes (a * b) + c, leveraging FMA if available
 #[inline]
+#[allow(clippy::suboptimal_flops)]
 pub fn multiply_add(a: f32, b: f32, c: f32) -> f32 {
     if cfg!(target_feature = "fma") {
         a.mul_add(b, c)