@FreezyLemon 2024-02-28
I'm actually not sure if it'd be worth using `multiply_add` over `mul_add` everywhere in the project. But the polyX functions seem to be the most important