@FreezyLemon 2024-02-28
I read up on libc::fmaf, and it seems like this change will result in reduced accuracy. `fmaf` seems to guarantee only one rounding error, but will use very slow code to achieve that on non-FMA platforms. Probably needs a bit more consideration