@shssoichiro 2024-02-29
Interesting. I'll accept this for now, but I think I actually want to upstream this to the Rust stdlib. It seems like a common footgun.