@FreezyLemon 2024-02-29
It's a somewhat common misconception; can't find the link now but some people have been discussing the advantages and disadvantages of mul_add in some GH issue (clippy I think? they have a lint that recommends mul_add over `a * b + c`) before.

The thing is that you shouldn't use `mul_add` for performance gains because the only thing it *guarantees* is the higher accuracy (one rounding error instead of two). The speed increase is more of an incidental thing, but it's reliable on any x86-64-v3 (or later) target.

So I guess it was our fault for using `mul_add` and expecting speed increases?