@shssoichiro 2024-02-29
It should certainly give improvements when built with FMA, because without `mul_add`, the compiler will not generate FMA instructions. The issue is only when compiling without FMA support.