@FreezyLemon 2024-02-29
Yeah that's what I meant. It's a bit weird because on the one hand, rustc doesn't seem to generate FMA instructions without mul_add. On the other hand, `mul_add`'s codegen seems to imply that the primary concern is the improved precision of the floating-point operation because it will favor high precision over fast codegen (mul_add *could* just be one mul and one add instruction if you don't care about performance)