@shssoichiro 2024-03-01
Not surprisingly, Rust won't fix this as it alters the result: https://github.com/rust-lang/rust/issues/112192

There is some more in-depth talk about a `-ffast-math` mode here: https://github.com/rust-lang/rust/issues/21690 but it doesn't sound like we're likely to get it any time soon, so it seems like we're stuck doing a manual workaround. Unfortunate.